+++ src/routes/IntegrationRoute.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { RouteComponentProps } from "react-router";
 
-import IntegrationPage from "./pages/integration/StateIntegrationPage";
+import IntegrationPage from "../pages/integration/StateIntegrationPage";
 
 interface IntegrationProps extends RouteComponentProps<any> {
